New class, ZSlider, that is used as a z level slider and text field

+++ ucar/unidata/idv/control/DrawingControl.java
@@ -1571,7 +1571,7 @@ public class DrawingControl extends DisplayControlImpl {
      *
      * @param value the new z slider value
      */
-    public void zSliderChanged(int value) {
+    public void zSliderChanged(double value) {
         super.zSliderChanged(value);
         if (displayOnly) {
             try {
